- fixed but in reporting download progress
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@1144 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/net/ThreadedURLDownload.java
@@ -96,8 +96,9 @@ public class ThreadedURLDownload implements Runnable {
                 totalread += bytesRead;
                 if (dd != null) {
                     if (totalsize != -1) {
-                        dd.setProgress((int)(100*(totalread/totalsize)));
-                        dd.setAmountDownloaded(MiscUtils.getSophisticatedByteString(totalread) + " downloaded (" + (100*(totalread/totalsize)) + " %)");
+
+                        dd.setProgress(((int) (100*(((double)totalread)/totalsize))));
+                        dd.setAmountDownloaded(MiscUtils.getSophisticatedByteString(totalread) + " downloaded (" +  ((int) (100*(((double)totalread)/totalsize))) + " %)");
                     } else {
 
                         dd.setAmountDownloaded(MiscUtils.getSophisticatedByteString(totalread) + " downloaded");
